Some fixes too
