+++ test/Microsoft.ML.Tests/Scenarios/Api/Estimators/Evaluation.cs
@@ -2,11 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.ML.Runtime.Api;
 using Microsoft.ML.Runtime.Data;
 using Microsoft.ML.Runtime.Learners;
 using Xunit;
-using Microsoft.ML.Models;
 
 namespace Microsoft.ML.Tests.Scenarios.Api
 {
