+++ test/Microsoft.ML.Tests/Scenarios/Api/Estimators/Extensibility.cs
@@ -1,10 +1,12 @@
-﻿using Microsoft.ML.Runtime.Api;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.ML.Runtime.Api;
 using Microsoft.ML.Runtime.Data;
 using Microsoft.ML.Runtime.Learners;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using Xunit;
 
 namespace Microsoft.ML.Tests.Scenarios.Api
