+++ test/Microsoft.ML.Tests/Scenarios/Api/Estimators/Visibility.cs
@@ -1,4 +1,8 @@
-﻿using Microsoft.ML.Runtime.Data;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.ML.Runtime.Data;
 using Xunit;
 
 namespace Microsoft.ML.Tests.Scenarios.Api
